refactor: mover EstadoCorrectoraSchema a entidades/estado.ts
`tenant/cliente.model.ts` necesita el schema como valor para un `z.partialRecord`
y tomarlo de `entidades/correctora.ts` cerraba un ciclo runtime real:
correctora -> localidad.ts -> ClienteSchema -> correctora (localidad.ts importa
ClienteSchema como valor, no como tipo).

`entidades/estado.ts` es un archivo hoja: solo depende de zod. `correctora.ts` no
re-exporta el schema porque el barrel de entidades ya exporta `./estado` y un
`export *` duplicado haria ambigua la exportacion.

Para los consumidores no cambia nada: importan de 'modelos' y el barrel sigue
exportando IEstado / EstadoCorrectoraSchema una sola vez.
