+++ src/interfaces/entidades/index.ts
@@ -14,6 +14,7 @@ export * from "./cuenca";
 export * from "./cuenta-cliente";
 export * from "./dispositivo";
 export * from "./envio-sms";
+export * from "./estado";
 export * from "./grupo";
 export * from "./kmz";
 export * from "./localidad";
