+++ src/interfaces/entidades/medidor-residencial.ts
@@ -2,7 +2,8 @@ import { z } from 'zod';
 import { CoordenadasSchema, ICoordenadas } from '../auxiliares/coordenadas';
 import { CentroOperativoSchema } from '../gas/centroOperativo/schema';
 import { UnidadNegocioSchema } from '../gas/unidadNegocio/schema';
-import { EstadoCorrectoraSchema, IEstado } from './correctora';
+import { EstadoCorrectoraSchema } from './estado';
+import type { IEstado } from './estado';
 import { CuencaSchema } from './cuenca';
 import { GrupoSchema } from './grupo';
 import { LocalidadSchema } from './localidad';
