+++ src/interfaces/entidades/scada.ts
@@ -2,7 +2,8 @@ import { z } from "zod";
 import { CentroOperativoSchema } from "../gas/centroOperativo/schema";
 import { UnidadNegocioSchema } from "../gas/unidadNegocio/schema";
 import { LocalidadSchema } from "./localidad";
-import { EstadoCorrectoraSchema, IEstado } from "./correctora";
+import { EstadoCorrectoraSchema } from "./estado";
+import type { IEstado } from "./estado";
 import { ClienteSchema } from "../tenant/cliente.model";
 import type { IAlerta } from "./alerta";
 import type { IConfigDispositivoScada } from "./config-dispositivo";
